change succes

+++ Assets/API/Engine/Card.cs
@@ -108,7 +108,7 @@ namespace Engine
         }
 
         //Constructor para las cartas creadas por el usuario
-        public Card(CardType type,string name, Faction faction, int points, Position position,Guid id)
+        public Card(CardType type,string name, Faction faction, double points, Position position,Guid id)
         {
             CardType = type;
             this.name = name;
